Refactor layout to GetX controller
