+++ lib/services/initial_bindings.dart
@@ -1,9 +1,11 @@
 import 'package:get/get.dart';
 import '../modules/auth/controllers/auth_controller.dart';
+import '../modules/layout/controllers/layout_controller.dart';
 
 class InitialBindings extends Bindings {
   @override
   void dependencies() {
     Get.put(AuthController());
+    Get.put(LayoutController());
   }
 }
